+++ foursight_core/environment.py
@@ -1,77 +1,174 @@
+import logging
+
 from .s3_connection import S3Connection
+from dcicutils.common import EnvName, ChaliceStage
+from dcicutils.env_manager import EnvManager
+from dcicutils.env_utils import (
+    get_foursight_bucket, get_foursight_bucket_prefix, full_env_name, infer_foursight_from_env,
+    get_env_real_url,
+)
+from dcicutils.misc_utils import full_class_name
+from dcicutils.s3_utils import s3Utils, HealthPageKey
+from typing import Optional, List
+
+
+logging.basicConfig()
+logger = logging.getLogger(__name__)
 
 
 class Environment(object):
 
-    def __init__(self, foursight_prefix):
-        self.prefix = foursight_prefix
+    def __init__(self, foursight_prefix: Optional[str] = None):  # the foursight_prefix argument can go away.
+
+        prefix = get_foursight_bucket_prefix()
+
+        # This consistency check AND INTERIM HEURISTIC ERROR CORRECTION can go away later,
+        # but when it does we can also get
+        # rid of the argument that is passed, since it should be possible to look up.
+        # For now we have made the argument optional so that it can be gradually phased out,
+        # but failing to pass it causes a requirement to declare the "foursight_bucket_prefix"
+        # in the ecosystem file of the global env bucket. -kmp 22-Jun-2022
+        if not prefix:
+            if foursight_prefix is not None:
+                prefix = foursight_prefix
+                logger.error(f"There is no declared foursight bucket prefix."
+                             f" The value from a deprecated data flow path, {foursight_prefix!r}, will be used.")
+            else:
+                raise RuntimeError(f"There is no declared foursight bucket prefix,"
+                                   f" and no suitable substitute can be inferred through legacy means.")
+        elif foursight_prefix and prefix != foursight_prefix:
+            logger.error(f"A value of deprecated argument 'foursight_prefix', {foursight_prefix},"
+                         f" was given when initializing an instance of the class {full_class_name(self)},"
+                         f" but the value does not match the declared foursight bucket prefix, {prefix}.")
+
+        self.prefix = prefix
         self.s3_connection = S3Connection(self.get_env_bucket_name())
 
-    def get_env_bucket_name(self):
-        return self.prefix + '-envs'
+    def get_env_bucket_name(self) -> Optional[str]:
 
-    def list_environment_names(self):
-        """
-        Lists all environments in the foursight-envs s3. Returns a list of names
-        """
-        return self.s3_connection.list_all_keys()
+        bucket_name = EnvManager.global_env_bucket_name()
 
-    def list_valid_schedule_environment_names(self):
-        """Lists all valid environ names used in schedules including 'all'"""
+        # This consistency check can go away later but has been downgraded to a warning after initial testing
+        # confirms this approach can work.
+        legacy_bucket_name = self.prefix + '-envs'
+        if bucket_name != legacy_bucket_name:
+            # Note the inconsistency, but don't flag an error.
+            logger.warning(f"{full_class_name(self)}.get_env_bucket_name is returning {bucket_name!r},"
+                           f" where it previously would have returned {legacy_bucket_name!r}.")
+
+        return bucket_name
+
+    def list_environment_names(self) -> List[EnvName]:
+        """
+        Lists all environments in the foursight-envs s3.
+
+        Returns: a list of names
+        """
+
+        environment_names = [infer_foursight_from_env(envname=env)
+                             for env in sorted(EnvManager.get_all_environments(env_bucket=self.get_env_bucket_name()))]
+
+        # This consistency check can go away later..
+        legacy_return_value = sorted([key
+                                      for key in self.s3_connection.list_all_keys()
+                                      if not key.endswith(".ecosystem")])
+        modern_full_names = list(map(full_env_name, environment_names))
+        legacy_full_names = list(map(full_env_name, legacy_return_value))
+        if modern_full_names != legacy_full_names:
+            logger.warning(f"{full_class_name(self)}.list_environment_names has consistency problems.")
+
+        return environment_names
+
+    def list_valid_schedule_environment_names(self) -> List[EnvName]:
+        """
+        Lists all valid environ names used in schedules including 'all'.
+
+        Returns: A list of names.
+        """
+
         return self.list_environment_names() + ['all']
 
-    def is_valid_environment_name(self, env):
-        """check if env is a valid environment name"""
+    def is_valid_environment_name(self, env: Optional[EnvName]) -> bool:
+        """
+        Returns True if env is a valid environment name, and False otherwise.
+
+        :param env: The name of an environment.
+        """
+
+        # TODO: Figure out whether callers mean this to return True for all possible forms of an environment
+        #       or only the actually-declared name. -kmp 22-Jun-2022
+
+        # This call requires no changes. -kmp 24-May-2022
         if env in self.list_environment_names():
             return True
         else:
             return False
 
-    def get_environment_info_from_s3(self, env_name):
-        return self.s3_connection.get_object(env_name)
+    @classmethod
+    def get_environment_info_from_s3(cls, env_name: EnvName) -> dict:
 
-    def get_environment_and_bucket_info(self, env_name, stage):
-        env_res = self.get_environment_info_from_s3(env_name)
-        # check that the keys we need are in the object
-        if isinstance(env_res, dict) and {'fourfront', 'es'} <= set(env_res):
-            env_entry = {
-                'fourfront': env_res['fourfront'],
-                'es': env_res['es'],
-                'ff_env': env_res.get('ff_env', ''.join(['fourfront-', env_name])),
-                'bucket': ''.join([self.prefix + '-', stage, '-', env_name])
-            }
-            return env_entry
-        else:
-            raise Exception(f'malformatted environment info on s3 for key {env_name}\n'
-                            f'{env_res}')
+        return s3Utils.get_synthetic_env_config(env_name)
 
-    def get_selected_environment_names(self, env_name):
+    def get_environment_and_bucket_info(self, env_name: EnvName, stage: ChaliceStage) -> dict:
+
+        env_info = self.get_environment_info_from_s3(env_name)
+
+        portal_url = env_info['fourfront']
+        es_url = env_info['es']
+        ff_env = env_info['ff_env']
+
+        bucket_name = get_foursight_bucket(envname=env_name, stage=stage)
+
+        # This consistency check can go away later. -kmp 22-Jun-2022
+        legacy_bucket_name = ''.join([self.prefix + '-', stage, '-', env_name])
+        if bucket_name != legacy_bucket_name:
+            logger.warning(f"{full_class_name(self)}.get_environment_and_bucket_info({env_name!r}, {stage!r})"
+                           f" is returning {bucket_name!r},"
+                           f" but it previously would have returned {legacy_bucket_name!r}.")
+
+        env_and_bucket_info = {
+            'fourfront': portal_url,
+            'es': es_url,
+            'ff_env': ff_env,
+            'bucket': legacy_bucket_name,
+        }
+        return env_and_bucket_info
+
+    def get_selected_environment_names(self, env_name: EnvName):
+
+        # This is weirdly named. A better name would be expand_environment_names or get_matching_environment_names.
+        # But it doesn't need to change. -kmp 24-May-2022
         if env_name == 'all':
             return self.list_environment_names()
         elif self.is_valid_environment_name(env_name):
             return [env_name]
         else:
-            raise Exception("not a valid env name")
+            raise ValueError(f"Not a valid env name: {env_name}")
 
-    def get_environment_and_bucket_info_in_batch(self, stage, env='all', envs=None):
+    def get_environment_and_bucket_info_in_batch(self, stage: ChaliceStage, env: Optional[EnvName] = None,
+                                                 envs: Optional[List[EnvName]] = None):
         """
         Generate environment information from the envs bucket in s3.
         Returns a dictionary keyed by environment name with value of a sub-dict
         with the fields needed to initiate a connection.
 
-        :param stage: the relevant chalice stage
+        :param stage: a chalice stage (generally one of 'dev' or 'prod')
+        :param env: allows you to specify a single env to be initialized, or the token 'all'.
         :param env: allows you to specify a single env to be initialized
         :param envs: allows you to specify multiple envs to be initialized
         """
         if envs is not None:
+            if env is not None:
+                ValueError(f"{full_class_name(self)}.get_environment_and_bucket_info_in_batch"
+                           f" accepts either 'env=' or 'envs=' but not both.")
             env_keys = envs
         else:
             try:
-                env_keys = self.get_selected_environment_names(env)
+                env_keys = self.get_selected_environment_names(env or 'all')
             except Exception:
                 return {}  # provided env is not in s3
 
-        environments = {}
-        for env_key in env_keys:
-            environments[env_key] = self.get_environment_and_bucket_info(env_key, stage)
-        return environments
+        return {
+            env_key: self.get_environment_and_bucket_info(env_key, stage)
+            for env_key in env_keys
+        }