+++ foursight_core/stage.py
@@ -1,4 +1,14 @@
+import boto3
+import logging
 import os
+import re
+
+from dcicutils.lang_utils import conjoined_list
+from dcicutils.cloudformation_utils import AbstractOrchestrationManager
+
+
+logging.basicConfig()
+logger = logging.getLogger(__name__)
 
 
 class Stage(object):
@@ -23,10 +33,23 @@ class Stage(object):
     def get_queue_name(self):
         return '-'.join([self.prefix, self.get_stage_from_env_variable(), 'check_queue'])
 
-    def get_runner_name(self):
+    ENCACHE_RUNNER_NAME = True
+
+    def  get_runner_name(self, encache=ENCACHE_RUNNER_NAME):
+        """
+        Gets the name of the Lambda function to use as a check runner.
+        """
+        stage = self.get_stage()
         check_runner = os.environ.get('CHECK_RUNNER', None)
+
         if not check_runner:
-            check_runner = '-'.join([self.prefix, self.get_stage(), 'check_runner'])
+            check_runner = AbstractOrchestrationManager.discover_foursight_check_runner_name(stage=stage,
+                                                                                             encache=encache)
+
+        if not check_runner:
+            # TODO: Is there EVER a situation where this works any more? Can we ditch this heuristic? -kmp 23-Jun-2022
+            check_runner = '-'.join([self.prefix, stage, 'check_runner'])
+
         return check_runner
 
     @classmethod